Defer metadata update up the folder tree to avoid massive contentions on parallel uploads
The patch is needed to solve the problem that the massive parallel uploads during migration produce a lot of LOCK contentions in MYSQL. The patch solves this by avoiding automatic tree update and providing an interface to launch update on demand.

This is not a planned NextCloud product feature yet and could be removed after migration (hopefully).

Signed-off-by: Bernd.Rederlechner@t-systems.com <bernd.rederlechner@t-systems.com>